+++ src/main/java/racingcar/domain/car/dto/input/CreateCarsRacing.java
@@ -1,7 +1,7 @@
 package racingcar.domain.car.dto.input;
 
 import java.util.List;
-import racingcar.domain.car.CarsValidator;
+import racingcar.domain.car.validator.CarsValidator;
 
 public record CreateCarsRacing(List<String> carNames) {
