+++ src/main/java/racingcar/domain/car/validator/CarsValidator.java
@@ -1,4 +1,4 @@
-package racingcar.domain.car;
+package racingcar.domain.car.validator;
 
 import java.util.List;
 import racingcar.exception.DuplicateCarNameException;
